% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calc_best_pattern.R
\name{calc_best_pattern}
\alias{calc_best_pattern}
\title{calculates the best fitting pattern}
\usage{
calc_best_pattern(
  Factors,
  targetVariable,
  dose,
  delta,
  order = seq_len(nrow(Factors)),
  compare_pattern
)
}
\arguments{
\item{Factors}{data frame with factors in columns}

\item{targetVariable}{vector with target variable (binary or metric)}

\item{dose}{vector with dose levels}

\item{delta}{vector with fuzzy logic threshold values}

\item{order}{order of the factors}

\item{compare_pattern}{pattern comparision (complement or overall means)}
}
\description{
calculates the best fitting pattern
}
