% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calc_truth_values.R
\name{calc_truth_values}
\alias{calc_truth_values}
\title{calculates truth values in dorisAutoPattern function}
\usage{
calc_truth_values(
  Factors,
  targetVariable,
  dose,
  delta,
  method = "Minimum",
  best_pattern_and_truthValues = levels_and_pattern,
  weights,
  compare_pattern
)
}
\arguments{
\item{Factors}{data frame with factors in columns}

\item{targetVariable}{vector with target variable (binary or metric)}

\item{dose}{vector with dose levels}

\item{delta}{vector with fuzzy logic threshold values}

\item{method}{method used to calculate best fitting pattern}

\item{best_pattern_and_truthValues}{best fitting patterns and truth values}

\item{weights}{weights for weigthes means}

\item{compare_pattern}{pattern comparision (complement or overall means)}
}
\description{
calculates truth values in dorisAutoPattern function
}
