% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doris_preprocessGraph.R
\name{doris_preprocessGraph}
\alias{doris_preprocessGraph}
\title{Pre-processing function for dorisGraph_base function}
\usage{
doris_preprocessGraph(
  Factors,
  dose,
  targetVariable,
  factor_selected,
  subgroup_selected,
  factor_selected2,
  subgroup_selected2,
  pattern,
  delta
)
}
\arguments{
\item{Factors}{data frame with factors in columns}

\item{dose}{vector with dose levels}

\item{targetVariable}{vector with target variable (binary or metric)}

\item{factor_selected}{first selected factor}

\item{subgroup_selected}{first selected subgroup}

\item{factor_selected2}{second selected factor}

\item{subgroup_selected2}{second selected subgroup}

\item{pattern}{selected pattern}

\item{delta}{vector with fuzzy logic threshold values}
}
\description{
Pre-processing function for dorisGraph_base function
}
