% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dorisAutoPattern.R
\name{dorisAutoPattern}
\alias{dorisAutoPattern}
\title{Calculates best fitting pattern}
\usage{
dorisAutoPattern(
  Factors,
  dose,
  targetVariable,
  delta,
  method = "Minimum",
  nperm = 1000,
  sampleseed = 1234,
  pattern_choice,
  pattern = NULL,
  perform_perm = FALSE,
  weights = NULL,
  compare_pattern
)
}
\arguments{
\item{Factors}{data frame with factors in columns}

\item{dose}{vector with dose levels}

\item{targetVariable}{vector with target variable (binary or metric)}

\item{delta}{vector with fuzzy logic threshold values}

\item{method}{method used to calculate best fitting pattern}

\item{nperm}{number of permutations}

\item{sampleseed}{seed for sampling in permutations}

\item{pattern_choice}{character, either "automatic" or "manual" pattern selection}

\item{pattern}{vector to specify pattern of differences}

\item{perform_perm}{logical, whether to perform permutation testing}

\item{weights}{numeric vector of weights for the "Weighted Mean" method}

\item{compare_pattern}{character, either "overall" or "subgroup complement"}
}
\description{
Calculates best fitting pattern
}
