% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dorisCalcPvalue.R
\name{dorisCalcPvalue}
\alias{dorisCalcPvalue}
\title{calculates pvalue for permutations in doris app}
\usage{
dorisCalcPvalue(tmp_list, truth_value)
}
\arguments{
\item{tmp_list}{reactive calc_permutation_automatic or calc_permutation_manual-object from doris app}

\item{total_truth_val}{numeric value of total truth value}
}
\description{
calculates pvalue for permutations in doris app
}
