% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dorisDistribution.R
\name{dorisDistribution}
\alias{dorisDistribution}
\title{calculates distribution of the number of subgroups identified as remarkable for a given trial with resampling}
\usage{
dorisDistribution(
  Factors,
  dose,
  targetVariable,
  pattern,
  delta,
  alpha = 0.05,
  runs
)
}
\arguments{
\item{Factors}{data frame with factors in columns}

\item{dose}{vector with dose levels}

\item{targetVariable}{vector with target variable (binary or metric)}

\item{pattern}{vector to specify pattern of differences}

\item{delta}{vector with fuzzy logic threshold values}

\item{alpha}{pattern matching threshold}

\item{runs}{number of simulation runs}
}
\value{
vector with frequencies of identified subgroups
}
\description{
calculates distribution of the number of subgroups identified as remarkable for a given trial with resampling
}
