% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dorisEvaluation.R
\name{dorisEvaluation}
\alias{dorisEvaluation}
\title{Function \code{dorisEvaluation} identifies remarkable subgroups.}
\usage{
dorisEvaluation(
  Factors,
  dose,
  targetVariable,
  pattern,
  delta,
  alpha = 0.05,
  method = "Minimum"
)
}
\arguments{
\item{Factors}{data frame with factors in columns}

\item{dose}{vector with dose levels}

\item{targetVariable}{vector with target variable (binary or metric)}

\item{pattern}{vector to specify pattern of differences}

\item{delta}{vector with fuzzy logic threshold values}

\item{alpha}{pattern matching threshold}
}
\value{
a list with six items (lowercase: vectors, uppercase: data frames):
* \code{ctm} clinical trial means
* \code{SGM} subgroup means
* \code{DSC} differences \code{SGM - ctm}
* \code{DTV} dose level truth values
* \code{stv} subgroup truth value
* \code{pmd} pattern matching decision
}
\description{
Function \code{dorisEvaluation} identifies remarkable subgroups.
}
