% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dorisGraph_base3.R
\name{dorisGraph_base3}
\alias{dorisGraph_base3}
\title{Main graph function in doris}
\usage{
dorisGraph_base3(
  dorisGraphData,
  lower,
  upper,
  subgroup,
  subgroup_level,
  compare_pattern,
  add_overall_mean_logical,
  add_complement_logical,
  add_other_subgroups_logical,
  add_backgrounds_logical,
  add_points_logical = FALSE,
  points_data,
  pattern_choice_auto = TRUE,
  add_permutation_infos = FALSE,
  jitter_points,
  tmp_list,
  index
)
}
\arguments{
\item{dorisGraphData}{data frame from preprocessing function}

\item{lower}{lower y-axis limit}

\item{upper}{upper y-axis limit}

\item{compare_pattern}{selected pattern}

\item{add_overall_mean_logical}{logical for adding mean lines}

\item{add_complement_logical}{logical for adding complement lines}

\item{add_other_subgroups_logical}{logical for adding other subgroups}

\item{add_backgrounds_logical}{logical for background}

\item{add_points_logical}{logical for adding points}

\item{points_data}{raw data points}

\item{pattern_choice_auto}{best fitting pattern}

\item{add_permutation_infos}{logical for permutation infos}

\item{jitter_points}{logical for jitter points}

\item{index}{}
}
\description{
Main graph function in doris
}
