% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doris_Histogram.R
\name{dorisHistogram}
\alias{dorisHistogram}
\title{Draws a histogram of permuted total truth values in doris app}
\usage{
dorisHistogram(tmp_list, total_truth_val, index)
}
\arguments{
\item{tmp_list}{reactive calc_permutation_automatic or calc_permutation_manual-object from doris app}

\item{total_truth_val}{numeric value of total truth value}

\item{index}{numeric index vector}
}
\description{
Draws a histogram of permuted total truth values in doris app
}
