% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dorisOverview.R
\name{dorisOverview}
\alias{dorisOverview}
\title{dose response in subgroups}
\usage{
dorisOverview(Factors, dose, targetVariable)
}
\arguments{
\item{Factors}{data frame with factors in columns}

\item{dose}{vector with dose levels}

\item{targetVariable}{vector with target variable (binary or metric)}
}
\value{
a list with three items (lowercase for vectors, uppercase for data frames):
* \code{fact} names of factors
* \code{levl} factor levels
* \code{SGDL} crosstab of subgroups by dose levels
}
\description{
Function \code{dorisOverview} provides an overview of the data
and the information to identify numbered subgroups, as provided
by dorisEvaluation. The "overview" is seperated from the "evaluation"
to keep \code{dorisEvaluation} efficient for simulations.
}
