% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utls_gettvpdose.R
\name{gettvpdose}
\alias{gettvpdose}
\title{helper function to get truth value per dose}
\usage{
gettvpdose(x, delta)
}
\arguments{
\item{x}{numeric vector}

\item{delta}{vector with fuzzy logic threshold values}
}
\description{
helper function to get truth value per dose
}
