% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpTextDropdownButton.R
\name{helpTextDropdownButton}
\alias{helpTextDropdownButton}
\title{helper function for helptext in doris app}
\usage{
helpTextDropdownButton(id, text, right)
}
\arguments{
\item{id}{identifier name}

\item{text}{help text}
}
\description{
helper function for helptext in doris app
}
