% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_doris.R
\name{run_doris}
\alias{run_doris}
\title{Run the Shiny Application}
\usage{
run_doris(host = NULL, port = NULL, ...)
}
\arguments{
\item{host}{host name (see [shiny::runApp()])}

\item{port}{port number (see [shiny::runApp()])}

\item{...}{A series of options to be used inside the app.}
}
\description{
Run the Shiny Application
}
