% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dorisAutoPattern.R
\name{dorisAutoPattern}
\alias{dorisAutoPattern}
\title{Calculates best fitting pattern}
\usage{
dorisAutoPattern(
  Factors,
  dose,
  targetVariable,
  delta,
  method = "Minimum",
  nperm = 1000,
  sampleseed = 1234,
  pattern_choice,
  pattern = NULL,
  perform_perm = FALSE,
  weights = NULL,
  compare_pattern
)
}
\arguments{
\item{Factors}{data frame with factors in columns}

\item{dose}{vector with dose levels}

\item{targetVariable}{vector with target variable (binary or metric)}

\item{delta}{vector with fuzzy logic threshold values}

\item{method}{method used to calculate best fitting pattern}

\item{pattern}{vector to specify pattern of differences}

\item{pattern_list}{vector with possible patterns in the form of e.g."=>>>="}
}
\description{
Calculates best fitting pattern
}
